Archive and clean Package Directory

This function processes an R package directory, removing irrelevant files (e.g., .git, .gitignore-listed files, etc.) and strips parts of lines matching a specified regex from source code. It then creates a clean folder or a zipped archive of the package without modifying the original directory.

MAJOR NOTE this is no where near perfect and the regexes used at the moment for removing comments are prone to issues such as not detecting whether the symbols are inside special strings.

Usage

archiveCleanedPackage(
  packageDir,
  output,
  removeHidden = TRUE,
  removeGitignore = TRUE,
  verbose = FALSE
)

Arguments

packageDir

A string specifying the path to the package directory.

output

A string specifying the name of the output cleaned directory. If the string ends with .zip, the function will create a zipped archive of the cleaned directory.

removeHidden

Logical; if TRUE, removes hidden files and folders (starting with a dot). Default is TRUE.

removeGitignore

Logical; if TRUE, removes files listed in the .gitignore. Default is TRUE.

verbose

Logical; if TRUE, prints additional information about the cleaning process. Default is FALSE.

Value

A cleaned directory or a zipped file of the package directory.

Details

The function cleans the package directory by:

Removing hidden files and folders like .git.

Excluding files listed in .gitignore.

Stripping out parts of lines matching a regex pattern from source files.

The output will be either a cleaned directory or a zipped archive of it.

Examples

if (FALSE) { # \dontrun{
# Clean and archive the package directory
archiveCleanedPackage("/path/to/package", "clean_package.zip")

# Clean package directory without zipping it
archiveCleanedPackage("/path/to/package", NULL)
} # }