List Only Files in a Directory

This function lists only the files in a specified directory, excluding directories. It is useful when you need to process or analyze only the files within a directory without including subdirectories. The base::list.files() function lists both files and directories, so this function provides a more convenient way to obtain just the files.

Usage

listFiles(dirPath, ...)

Arguments

dirPath

Character. The path to the directory from which to list files.

...

Additional arguments passed to base::list.files() (e.g., pattern, recursive). Note that full.names will be ignored.

Value

A character vector of file paths.

Examples

# \donttest{
listFiles(getwd())
#>  [1] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/FastUtils-package.html"        
#>  [2] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/add.html"                      
#>  [3] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/bound.html"                    
#>  [4] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/closestWord.html"              
#>  [5] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/colToRownames.html"            
#>  [6] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/createHash.html"               
#>  [7] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/createMutator.html"            
#>  [8] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/createPkgLoader.html"          
#>  [9] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/divide.html"                   
#> [10] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/enclose.html"                  
#> [11] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/encloseBr.html"                
#> [12] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/enumerateit.html"              
#> [13] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/evalText.html"                 
#> [14] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/findMissingRdSections.html"    
#> [15] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/fixColnames.html"              
#> [16] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getAvgHex.html"                
#> [17] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getChar.html"                  
#> [18] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getFailStr.html"               
#> [19] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getPkgKeywords.html"           
#> [20] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getPlotDims.html"              
#> [21] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getUniquePairsUpTo.html"       
#> [22] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getfirst.html"                 
#> [23] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/getlast.html"                  
#> [24] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/greplDir.html"                 
#> [25] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/ind.html"                      
#> [26] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/index.html"                    
#> [27] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/initDataFrameWithColnames.html"
#> [28] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/initEmptyTable.html"           
#> [29] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/initList.html"                 
#> [30] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/initTestthat.html"             
#> [31] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/initV.html"                    
#> [32] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isBound.html"                  
#> [33] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isCamelCase.html"              
#> [34] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isEven.html"                   
#> [35] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isOdd.html"                    
#> [36] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isPascalCase.html"             
#> [37] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isSnakeCase.html"              
#> [38] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/isVowel.html"                  
#> [39] "/home/runner/work/FastUtils/FastUtils/docs/dev/reference/joinRegex.html"                
listFiles(getwd(), pattern = "\\.R$", recursive = TRUE)
#> character(0)
# }